requirements.txt
pandas
scikit-learn
statsmodels
mysql-connector-python
lxml
matplotlib
pymc3
arviz
ipykernel
